# fixes #24570 - dynamicailly import password strength component