--- a/package.json
+++ b/package.json
@@ -47,6 +47,7 @@
     "prettier": "^1.13.5",
     "raf": "^3.4.0",
     "raw-loader": "^0.5.1",
+    "react-loadable": "^5.5.0",
     "react-redux-test-utils": "^0.1.1",
     "react-remarkable": "^1.1.3",
     "react-test-renderer": "^16.8.1",