… webpack/assets/javascripts/react_app/components/PasswordStrength/__tests__/__snapshots__/integration.test.js.snap