this reduced the vendor file by about 800kb and reduce
every full page reload by about 100ms on my machine