@ohadlevy 2019-04-29
webpack analyze before
![image](https://user-images.githubusercontent.com/20167/56926794-3751ca80-6adb-11e9-84ba-9a1df1c5dce1.png)
webpack analyze after
![image](https://user-images.githubusercontent.com/20167/56926852-53556c00-6adb-11e9-8224-6bb7edc202e0.png)

vendor file size decreased by about 800K 
(not to mention that it should not be exposed to plugins to begin with imho)

and my chrome loading statistics shows a decrease in script evaluation for about 85ms +-