@ares 2019-04-30
Could you please help me understand how the loading changed? Is this loading the component somehow via async request only when needed?