@ohadlevy 2019-04-30
I forgot to mention this also uses the patternfly loaderState component, that only shows a spinner if its loading over 250ms (or something like that),