@ohadlevy 2019-04-30
> Not going to comment on the implementation, but the general direction sounds correct to me. 

I'm looking for input from @sharvit and @amirfefer etc... 
> Will this be an issue for plugins or is it fully transparent for them?

I don't expect any issues.