@ohadlevy 2019-05-26
OK, I've a few updates for this thread, I've

Added Spice 
Added Editor (which also bundle patternfly-react-extensions)

I've done the following performance tests

tl;dr


  | Current | Proposed | Difference (ms)
-- | -- | -- | --
First Load | 1677 | 1127 | 550
Average Page Load | 711.4 | 641.8 | 69.6


Memory size impact is about half on initial page loads


Bundle size impact

  | Current | Proposed
-- | -- | --
vendor | 4.57 MB | 2.31 MB
bundle | 896.66 KB | 488.99 KB
spice |   | 141.53 KB
noVNC |   | 138.6 KB
Editor |   | 1.57 MB
passwordStrength |   | 822.92 KB


I've done a total of 7 tests for each environment (as its done manually via chrome performance developers tools, raw data at


Current | ms | msLoading | msScripting | msRendering | msPainting | msSystem | msIdle
-- | -- | -- | -- | -- | -- | -- | --
First Load | 1677 | 49.3 | 1259.4 | 34.7 | 1.1 | 191.6 | 141.2
2 | 650 | 36.1 | 406.1 | 17 | 0.6 | 132.5 | 58.2
3 | 640 | 33.4 | 390.7 | 21.3 | 0.6 | 135.7 | 58.6
4 | 767 | 37.1 | 447.7 | 21.8 | 5.4 | 174.1 | 80.9
5 | 705 | 38 | 365.7 | 20.4 | 0.9 | 213.6 | 66.7
6 | 795 | 30.1 | 378.4 | 16 | 0.7 | 296.7 | 72.9
Proposed |   |   |   |   |   |   |  
First Load | 1127 | 32.7 | 870.5 | 46.4 | 0.9 | 73 | 103.2
2 | 538 | 28.9 | 315.5 | 16.9 | 0.6 | 87.4 | 88.5
3 | 770 | 8.1 | 291.4 | 14.3 | 0.9 | 118.9 | 336.3
4 | 485 | 11.7 | 297.6 | 15 | 0.6 | 90.6 | 69.7
5 | 653 | 6.3 | 290.9 | 15.1 | 0.7 | 109 | 230.7
6 | 763 | 8.1 | 298.1 | 16.7 | 0.8 | 101.3 | 338.2



If you would like to try this out, simply change your docker-compose.yml file
```diff
diff --git a/docker-compose.yml b/docker-compose.yml
index 821984ff2..16d67676e 100644
--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -20,7 +20,8 @@ services:
       - db:/var/lib/mysql/data
 
   app: &app_base
-    image: quay.io/foreman/foreman:develop
+    image: quay.io/ohadlevy/foreman:react-loadable
     command: bundle exec bin/rails server -b 0.0.0.0
     build:
       context: .
```