@ohadlevy 2019-05-28
@glekner can you have a look if you can find a way to get the integration tests working?