@ohadlevy 2019-06-10
@sharvit https://stackoverflow.com/questions/50176706/how-test-a-react-loadable-component has some options, what do you think?