@sharvit 2019-09-23
I am sorry for the inconvenient @ohadlevy but this PR is not mergeable because it should be implemented in the vendor now.

Adding it to the `foreman-js` roadmap.md as it should be there:
https://github.com/theforeman/foreman-js/pull/50